cs340-project/requirements.txt
blinker==1.9.0
click==8.1.7
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.6
platformdirs==4.3.6
python-dotenv==1.0.1
virtualenv==20.27.1
Werkzeug==3.1.3
